Lab2_Worker/requirements.txt
psycopg2-binary==2.9.3
Flask==2.0.1
Flask-Migrate==3.1.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
